requirements.consumer.txt
boto3
confluent-kafka
python-dotenv
stripe
google-cloud-bigquery
playwright